delete hotovo
treba ho ale dva krat stlacit aby sa item vymazal

+++ support.php
@@ -79,14 +79,27 @@ if (isset($_POST['title'])) {
                 <h3><?php echo $article->getTitle() ?></h3>
                 <p><?php echo $article->getText() ?></p>
                 <p><?php echo $article->getId() ?></p>
+                <a href="?fn=selection&id=<?= $article->getId() ?>">Zmazat</a>
 
                 <?php
-                $id = 1;
-                $storage->Delete($id);
+                if ($_GET['fn'] == "selection") {
+                    if(isset($_GET['id'])){
+
+                        $id = $_GET['id'];
+                        $storage->Delete($id);
+                    }
+                }
                 ?>
             </div>
         </div>
     </div>
 <?php } ?>
+
+
+
+
+
+
 </body>
 </html>
+
